Store Support/Projects/Intake Hub/Intake Hub/ProjectsinStores/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
google-cloud-bigquery==3.14.1
google-auth==2.27.0
openai==1.10.0
aiohttp==3.9.1
python-multipart==0.0.6

# Email Reporting Dependencies
apscheduler==3.10.4
reportlab==4.0.7
email-validator==2.1.0
